Capstone 3/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "uqishafi",email = "uqishafi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <4.0.0"
dependencies = [
    "langchain (>=1.2.0,<2.0.0)",
    "langchain-qdrant (>=1.1.0,<2.0.0)",
    "langchain-openai (>=1.1.5,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "langfuse (>=3.11.0,<4.0.0)",
    "langchain-classic (>=1.0.0,<2.0.0)",
    "streamlit (>=1.52.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "agent", from = "modules"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
